feat: adopt candor-tone-picker, remove _candor, refresh docs (stages 4-5)
Completes #149.

cc-tone-picker becomes candor-tone-picker. The two were near-identical — same
props, same cell shape, same private helpers — so the swap is attribute names
(hideHeaders -> hide-headers, selectedValue -> selected-value, ariaLabel ->
aria-label) and types (GridRow/GridCell -> ToneRow/ToneCell). The upstream
color-select event carries the coordinates alongside the string, so
palette-table no longer re-parses the oklch() it just received.

cc-table stays. candor-table is data-driven -- headers: string[], rows: {
cells: string[] }[] -- and half the tables in cc-metadata put an interactive
tooltip and info button inside a cell. There is nowhere in that API to put
one, and passing a TemplateResult through a string[] would work only until
upstream decided to coerce it. So cc-table moves to _components/ where an
app-owned light-DOM styling wrapper belongs, drops its .table-host class hook
in favour of a reflected `compact` attribute, and src/app/_candor/ is gone.

Docs: the Candor sections of architecture.md, components.md and
contributing.md described a directory that no longer exists, so they are
rewritten around the package -- including what you can and cannot restyle
across a shadow boundary, and why the cc-*/candor-* prefixes are load-bearing.
CLAUDE.md's Angular conventions section is replaced with Lit ones; the app
stopped being Angular in 36bd8e2. README's stack and structure follow. The
rest of documentation/ is still Angular-era prose and keeps its warning
banner.

Verified against a production build: the tone picker renders 10 rows with the
right labels, activating a cell round-trips through palette-table to update
the foreground color, and all four metadata tables render with the reflected
compact attribute.
