--- src/app/_components/tone-picker/tone-picker.component.scss
@@ -1,140 +1,0 @@
-cc-tone-picker {
-  display: block;
-  font-family: var(--font-family-accessible);
-}
-
-cc-tone-picker .gamut-grid {
-  border-collapse: collapse;
-  width: auto;
-
-  th,
-  td {
-    border: 0;
-    background: none;
-  }
-}
-
-cc-tone-picker .corner {
-  padding: 0;
-}
-
-.col-header,
-cc-tone-picker .row-header {
-  font-size: var(--font-size-xs);
-  font-weight: 400;
-  color: var(--color-text-subtle);
-  letter-spacing: 0.02em;
-}
-
-cc-tone-picker .col-header {
-  padding: 0.25rem 0.5rem;
-  text-align: center;
-  white-space: nowrap;
-}
-
-cc-tone-picker .row-header {
-  padding: 0.25rem 0.75rem 0.25rem 0;
-  text-align: right;
-  white-space: nowrap;
-}
-
-cc-tone-picker .cell {
-  // Padding must be >= outline-width + outline-offset (2px + 2px = 4px) to avoid clipping the focus ring.
-  padding: 4px;
-  width: 3.125rem;
-  height: 3.125rem;
-}
-
-cc-tone-picker .size-small .cell {
-  width: 1.625rem;
-  height: 1.625rem;
-  padding: 4px;
-}
-
-cc-tone-picker .cell-btn {
-  display: block;
-  width: 100%;
-  height: 100%;
-  border: 2px solid transparent;
-  border-radius: var(--radius-sm);
-  cursor: pointer;
-  padding: 0;
-
-  &:focus-visible {
-    outline: 2px solid var(--color-focus);
-    outline-offset: 2px;
-  }
-
-  &[aria-checked='true'] {
-    border-color: var(--color-focus);
-    outline: 2px solid var(--color-focus);
-    outline-offset: 2px;
-  }
-}
-
-// ─── Visually hidden header variant ──────────────────────────────────────────
-// Headers stay in the DOM so screen readers retain column/row context in the grid.
-
-cc-tone-picker .gamut-grid.hide-headers {
-  thead {
-    position: absolute;
-    width: 1px;
-    height: 1px;
-    padding: 0;
-    overflow: hidden;
-    clip: rect(0, 0, 0, 0);
-    clip-path: inset(50%);
-    white-space: nowrap;
-  }
-
-  .row-header,
-  .corner {
-    width: 0;
-    min-width: 0;
-    max-width: 0;
-    padding: 0;
-    overflow: hidden;
-    white-space: nowrap;
-    clip-path: inset(50%);
-  }
-}
-
-cc-tone-picker .preview {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  margin-top: 1.25rem;
-  min-height: 2rem;
-}
-
-cc-tone-picker .preview-swatch {
-  display: inline-block;
-  width: 2rem;
-  height: 2rem;
-  border-radius: var(--radius-sm);
-  flex-shrink: 0;
-  border: 1px solid var(--color-border-default);
-}
-
-cc-tone-picker .preview-code {
-  font-family: var(--font-family-mono);
-  font-size: var(--font-size-sm);
-  color: var(--color-text-default);
-  background: var(--color-bg-surface);
-  padding: 0.2em 0.5em;
-  border-radius: var(--radius-sm);
-  letter-spacing: 0.02em;
-}
-
-cc-tone-picker .preview-empty {
-  font-size: var(--font-size-sm);
-  color: var(--color-text-subtle);
-  letter-spacing: 0.02em;
-}
-
-cc-tone-picker .hint {
-  margin-top: 0.75rem;
-  font-size: var(--font-size-xs);
-  color: var(--color-text-subtle);
-  letter-spacing: 0.02em;
-}
