+++ src/styles.scss
@@ -7,7 +7,6 @@ cc-color-slider,
 cc-copy-to-clipboard-button,
 cc-metadata,
 cc-palette-table,
-cc-tone-picker,
 cc-table {
   display: block;
 }
